Removed unused metrics types
