--- metrics/doc.go
@@ -1,3 +1,0 @@
-package metrics
-
-const epsilon = 0.0000000000000001
