--- metrics/gauge.go
@@ -1,10 +1,0 @@
-package metrics
-
-// Gauges hold an int64 value that can be set arbitrarily.
-type Gauge interface {
-	Snapshot() Gauge
-	Update(int64)
-	Dec(int64)
-	Inc(int64)
-	Value() int64
-}
