--- metrics/gauge_float64.go
@@ -1,8 +1,0 @@
-package metrics
-
-// GaugeFloat64s hold a float64 value that can be set arbitrarily.
-type GaugeFloat64 interface {
-	Snapshot() GaugeFloat64
-	Update(float64)
-	Value() float64
-}
